+++ server/test/klite/UtilsTest.kt
@@ -17,10 +17,10 @@ class UtilsTest {
   }
 
   @Test fun plus() {
-    expect(URI("http://codeborne.com") + mapOf("hello" to "world 123")).toEqual(URI("http://codeborne.com?hello=world+123"))
-    expect(URI("http://codeborne.com?world=hello") + mapOf("hello" to "world 123")).toEqual(URI("http://codeborne.com?world=hello&hello=world+123"))
-    expect(URI("https://codeborne.com/#ref") + mapOf("hello" to "world")).toEqual(URI("https://codeborne.com/?hello=world#ref"))
-    expect(URI("https://codeborne.com/?x=y#ref") + mapOf("hello" to "world")).toEqual(URI("https://codeborne.com/?x=y&hello=world#ref"))
+    expect(URI("http://example.com") + mapOf("hello" to "world 123")).toEqual(URI("http://example.com?hello=world+123"))
+    expect(URI("http://example.com?world=hello") + mapOf("hello" to "world 123")).toEqual(URI("http://example.com?world=hello&hello=world+123"))
+    expect(URI("https://example.com/#ref") + mapOf("hello" to "world")).toEqual(URI("https://example.com/?hello=world#ref"))
+    expect(URI("https://example.com/?x=y#ref") + mapOf("hello" to "world")).toEqual(URI("https://example.com/?x=y&hello=world#ref"))
   }
 
   @Test fun urlParams() {