+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 * jdbc: PooledDataSource.queryTimeout can be configured separately from pool connection timeout
 * server: MultipartParser and MultipartRenderer now support binary data/files
 * server: nullable List of @QueryParam will now get null value if parameter is missing, not empty list
-* server: new experimental metrics() handler
+* server: introduced Metrics for publishing of app metrics, and a metrics() handler
+* core: new extensions for easier working with Java HttpClient, like get()/post()/put()
 * xml: new module for fast and lightweight XML parsing
 * Java 25 support